It is simply defined as `f64` across every platform right now.

A use case hasn't been presented for a `float` type defined as the
highest precision floating point type implemented in hardware on the
platform. Performance-wise, using the smallest precision correct for the
use case greatly saves on cache space and allows for fitting more
numbers into SSE/AVX registers.

If there was a use case, this could be implemented as simply a type
alias or a struct thanks to `#[cfg(...)]`.

Closes #6592

The mailing list thread, for reference:

https://mail.mozilla.org/pipermail/rust-dev/2013-July/004632.html

+++ src/libstd/unstable/extfmt.rs
@@ -480,7 +480,7 @@ pub mod ct {
 #[doc(hidden)]
 #[allow(non_uppercase_statics)]
 pub mod rt {
-    use float;
+    use f64;
     use str;
     use sys;
     use num;
@@ -563,10 +563,10 @@ pub mod rt {
         };
         pad(cv, unpadded, None, PadNozero, buf);
     }
-    pub fn conv_float(cv: Conv, f: float, buf: &mut ~str) {
+    pub fn conv_float(cv: Conv, f: f64, buf: &mut ~str) {
         let (to_str, digits) = match cv.precision {
-              CountIs(c) => (float::to_str_exact, c as uint),
-              CountImplied => (float::to_str_digits, 6u)
+              CountIs(c) => (f64::to_str_exact, c as uint),
+              CountImplied => (f64::to_str_digits, 6u)
         };
         let s = to_str(f, digits);
         let head = if 0.0 <= f {
